문제스펙에 맞추어 생성자를 변경하라

+++ src/main/java/algorythm/Pro_1_2_16.java
@@ -31,7 +31,7 @@ class Pro_1_2_16 {
      * @param numerator 분자
      * @param denominator 분모
      */
-    Rational(long numerator, long denominator) {
+    Rational(int numerator, int denominator) {
       long gcd = gcd(numerator, denominator);
       this.numerator = numerator / gcd;
       this.denominator = denominator / gcd;
@@ -56,15 +56,17 @@ class Pro_1_2_16 {
      */
     private Rational calculate(Rational b, LongBinaryOperator operator) {
       if (this.denominator == b.denominator) {
-        return new Rational(operator.applyAsLong(this.numerator, b.numerator), denominator);
+        return new Rational(
+            (int) operator.applyAsLong(this.numerator, b.numerator),
+            (int) denominator);
       }
       long lcm = lcm(this.denominator, b.denominator);
       return new Rational(
-          operator.applyAsLong(
+          (int) operator.applyAsLong(
               this.numerator * (lcm / this.denominator),
               b.numerator * (lcm / b.denominator)
           ),
-          lcm
+          (int) lcm
       );
     }
 
@@ -73,8 +75,8 @@ class Pro_1_2_16 {
      */
     Rational times(Rational b) {
       return new Rational(
-          this.numerator * b.numerator,
-          this.denominator * b.denominator
+          (int) (this.numerator * b.numerator),
+          (int) (this.denominator * b.denominator)
       );
     }
 
@@ -82,7 +84,7 @@ class Pro_1_2_16 {
      * this / b 의 결과를 리턴한다.
      */
     Rational divideBy(Rational b) {
-      return this.times(new Rational(b.denominator, b.getNumerator()));
+      return this.times(new Rational((int) b.denominator, (int) b.getNumerator()));
     }
 
     /**
